Added two columns to the list view for OC Recommendations: number of recommendations and OC Cache index, a calculated value indicating how often the cache was recommended in relation to the number of logs. Modified the index in CacheHolder from Float to int - this is not really a restriction but we don't really want decimals in the calculated OC cache index.

+++ trunk/src/CacheWolf/CacheHolder.java
@@ -80,8 +80,8 @@ public int noFindLogs = 0;
 public int numRecommended = 0;
 /** Number of logs since start of recommendations system */
 public int numLogsSinceRecommendation = 0;
-/** Recommendation score: calculated as rations  numRecommended / numLogsSinceRecommendation  */
-public float recommendationScore = 0;
+/** Recommendation score: calculated as ratio  numRecommended / numLogsSinceRecommendation * 100 */
+public int recommendationScore = 0;
 /** True if this cache has travelbugs */
 public boolean has_bug = false;
 /** True if the cache description is stored in HTML format */
